Move App.css and index.css to ./css
